Produces sorted sgr file with third column corresponding to tag start (1) or tag end (0).
To be used with corresponding gff2wig genome revision file (e.g. gff2wig_mm9.pl)

All the "end" positions of the tags located on the (+) strand are assigned a value corresponding to "start" plus the gel cut size, and vice versa for the (-) strand. This method may give rise to negative "start" positions, they are set to 1 in such a case. The extended tags thus correspond to the initial fragments. This script thus performs the first step towards the transition to "binned" data, like the wig format, according to the following logic: from the "start" and "end" positions, two entries are created per extended fragment, an entry with the "Start" with which a value of 1 is associated, and conversely another entry with the "end" with which a value of -1 is associated. All the fragments are then ordered by chromosome and by "start" position.

Used in 

Cauchy P, Maqbool MA, Zacarias-Cabeza J, Vanhille L, Koch F, Fenouil R, Gut M, Gut I, Santana MA, Griffon A, Imbert J, Moraes-Cabé C, Bories JC, Ferrier P, Spicuglia S, Andrau JC. Nucleic Acids Res. 2016 May 5;44(8):3567-85

For some external files