Produces wig file from sorted sgr file having extended read start/end information as 3d column (1=start, 0=end)
mm9 version

This Perl script comes after the bedOrderExtendTags function. It makes it possible to establish an average score per base pair per genomic coordinate interval (binning) of reads extended over a given genome. A score variable is incremented by 1 and -1 respectively when a "start" or "end" is encountered. The score per interval is then divided by the size of the interval, specified as input. This technique is furthermore not disturbed by the occasional asymmetry between the number of fragments sequenced on Watson and Crick strands. When no fragment is encountered in a given interval, the value per base pair is therefore 0. From the first to the last chromosome of the given genome, a counter indicating the chromosomal position is incremented by the size of the specified window. as input and reset when it exceeds the length of the chromosome being processed. The result is a wig file with a value per line, corresponding to the score per base pair per interval calculated by reading the fragments.

Used in 

Cauchy P, Maqbool MA, Zacarias-Cabeza J, Vanhille L, Koch F, Fenouil R, Gut M, Gut I, Santana MA, Griffon A, Imbert J, Moraes-Cabé C, Bories JC, Ferrier P, Spicuglia S, Andrau JC. Nucleic Acids Res. 2016 May 5;44(8):3567-85

For some external files